Migrate test_mod_parsing.py onto explicit, catalog-verified hashes
Replaces parsed_mod_from_arc (auto-discovery of every .mod found under
--arcdir, now removed along with its dispatch branch) with a
parsed_mod fixture that resolves committed hashes against a shared
game_fs_root, reusing the same re5 files already verified in R2 via
test_mod_serialization.py's dataset.

Extracts game_fs_root/_game_dirs/R2_PROTOCOL_PREFIX out of
test_mod_serialization.py into tests/mtfw/conftest.py so both files
share one MTFW_FS mount per app_id per session instead of each
maintaining its own copy.
