+++ tests/mtfw/datasets/mod_parsing_hashes.json
@@ -1,0 +1,10 @@
+[
+    {
+        "app_id": "re5",
+        "mod_path_hash": "5d45d4682b062d49"
+    },
+    {
+        "app_id": "re5",
+        "mod_path_hash": "2dc02b22175deb2f"
+    }
+]
